+++ include/extractor/extraction_node.hpp
@@ -1,7 +1,8 @@
 #ifndef EXTRACTION_NODE_HPP
 #define EXTRACTION_NODE_HPP
 
-#include "traffic_lights.hpp"
+#include "traffic_flow_control_nodes.hpp"
+#include <cstdint>
 
 namespace osrm
 {
@@ -10,14 +11,19 @@ namespace extractor
 
 struct ExtractionNode
 {
-    ExtractionNode() : traffic_lights(TrafficLightClass::NONE), barrier(false) {}
+    ExtractionNode() : traffic_lights(TrafficFlowControlNodeDirection::NONE), barrier(false) {}
     void clear()
     {
-        traffic_lights = TrafficLightClass::NONE;
+        traffic_lights = TrafficFlowControlNodeDirection::NONE;
+        stop_sign = TrafficFlowControlNodeDirection::NONE;
+        give_way = TrafficFlowControlNodeDirection::NONE;
         barrier = false;
     }
-    TrafficLightClass::Direction traffic_lights;
+    TrafficFlowControlNodeDirection traffic_lights;
     bool barrier;
+
+    TrafficFlowControlNodeDirection stop_sign;
+    TrafficFlowControlNodeDirection give_way;
 };
 } // namespace extractor
 } // namespace osrm