+++ src/extractor/extraction_containers.cpp
@@ -5,6 +5,7 @@
 #include "extractor/name_table.hpp"
 #include "extractor/restriction.hpp"
 #include "extractor/serialization.hpp"
+#include "extractor/traffic_flow_control_nodes.hpp"
 #include "util/coordinate_calculation.hpp"
 #include "util/integer_range.hpp"
 
@@ -412,12 +413,17 @@ void ExtractionContainers::PrepareData(ScriptingEnvironment &scripting_environme
 {
     const auto restriction_ways = IdentifyRestrictionWays();
     const auto maneuver_override_ways = IdentifyManeuverOverrideWays();
-    const auto traffic_signals = IdentifyTrafficSignals();
+    const auto traffic_signals = IdentifyTrafficFlowControlNodes(external_traffic_signals);
+    const auto stop_signs = IdentifyTrafficFlowControlNodes(external_stop_signs);
+    const auto give_ways = IdentifyTrafficFlowControlNodes(external_give_ways);
 
     PrepareNodes();
     PrepareEdges(scripting_environment);
 
-    PrepareTrafficSignals(traffic_signals);
+    PrepareTrafficFlowControlNodes(traffic_signals, internal_traffic_signals);
+    PrepareTrafficFlowControlNodes(stop_signs, internal_stop_signs);
+    PrepareTrafficFlowControlNodes(give_ways, internal_give_way_signs);
+
     PrepareManeuverOverrides(maneuver_override_ways);
     PrepareRestrictions(restriction_ways);
     WriteCharData(name_file_name);
@@ -935,23 +941,23 @@ ExtractionContainers::ReferencedWays ExtractionContainers::IdentifyManeuverOverr
     return maneuver_override_ways;
 }
 
-void ExtractionContainers::PrepareTrafficSignals(
-    const ExtractionContainers::ReferencedTrafficSignals &referenced_traffic_signals)
+void ExtractionContainers::PrepareTrafficFlowControlNodes(
+    const ReferencedTrafficFlowControlNodes &referenced_traffic_control_nodes,
+    TrafficFlowControlNodes &internal_traffic_control_nodes)
 {
-    const auto &bidirectional_signal_nodes = referenced_traffic_signals.first;
-    const auto &unidirectional_signal_segments = referenced_traffic_signals.second;
+    const auto &bidirectional_traffic_control_nodes = referenced_traffic_control_nodes.first;
+    const auto &unidirectional_node_segments = referenced_traffic_control_nodes.second;
 
     util::UnbufferedLog log;
-    log << "Preparing traffic light signals for " << bidirectional_signal_nodes.size()
-        << " bidirectional, " << unidirectional_signal_segments.size()
-        << " unidirectional nodes ...";
+    log << "Preparing traffic control nodes for " << bidirectional_traffic_control_nodes.size()
+        << " bidirectional, " << unidirectional_node_segments.size() << " unidirectional nodes ...";
     TIMER_START(prepare_traffic_signals);
 
     std::unordered_set<NodeID> bidirectional;
     std::unordered_set<std::pair<NodeID, NodeID>, boost::hash<std::pair<NodeID, NodeID>>>
         unidirectional;
 
-    for (const auto &osm_node : bidirectional_signal_nodes)
+    for (const auto &osm_node : bidirectional_traffic_control_nodes)
     {
         const auto node_id = mapExternalToInternalNodeID(
             used_node_id_list.begin(), used_node_id_list.end(), osm_node);
@@ -960,7 +966,7 @@ void ExtractionContainers::PrepareTrafficSignals(
             bidirectional.insert(node_id);
         }
     }
-    for (const auto &to_from : unidirectional_signal_segments)
+    for (const auto &to_from : unidirectional_node_segments)
     {
         const auto to_node_id = mapExternalToInternalNodeID(
             used_node_id_list.begin(), used_node_id_list.end(), to_from.first);
@@ -972,8 +978,8 @@ void ExtractionContainers::PrepareTrafficSignals(
         }
     }
 
-    internal_traffic_signals.bidirectional_nodes = std::move(bidirectional);
-    internal_traffic_signals.unidirectional_segments = std::move(unidirectional);
+    internal_traffic_control_nodes.bidirectional_nodes = std::move(bidirectional);
+    internal_traffic_control_nodes.unidirectional_segments = std::move(unidirectional);
 
     TIMER_STOP(prepare_traffic_signals);
     log << "ok, after " << TIMER_SEC(prepare_traffic_signals) << "s";
@@ -1159,37 +1165,39 @@ ExtractionContainers::ReferencedWays ExtractionContainers::IdentifyRestrictionWa
     return restriction_ways;
 }
 
-ExtractionContainers::ReferencedTrafficSignals ExtractionContainers::IdentifyTrafficSignals()
+ExtractionContainers::ReferencedTrafficFlowControlNodes
+ExtractionContainers::IdentifyTrafficFlowControlNodes(
+    const std::vector<InputTrafficFlowControlNode> &external_nodes)
 {
     util::UnbufferedLog log;
-    log << "Collecting traffic signal information on " << external_traffic_signals.size()
-        << " signals...";
-    TIMER_START(identify_traffic_signals);
+    log << "Collecting traffic nodes information on " << external_nodes.size() << " nodes...";
+    TIMER_START(identify_traffic_flow_control_nodes);
 
     // Temporary store for nodes containing a unidirectional signal.
-    std::unordered_map<OSMNodeID, TrafficLightClass::Direction> unidirectional_signals;
+    std::unordered_map<OSMNodeID, TrafficFlowControlNodeDirection> unidirectional_traffic_nodes;
 
     // For each node that has a unidirectional traffic signal, we store the node(s)
     // that lead up to the signal.
-    std::unordered_multimap<OSMNodeID, OSMNodeID> signal_segments;
+    std::unordered_multimap<OSMNodeID, OSMNodeID> node_segments;
 
-    std::unordered_set<OSMNodeID> bidirectional_signals;
+    std::unordered_set<OSMNodeID> bidirectional_traffic_nodes;
 
-    const auto mark_signals = [&](auto const &traffic_signal) {
-        if (traffic_signal.second == TrafficLightClass::DIRECTION_FORWARD ||
-            traffic_signal.second == TrafficLightClass::DIRECTION_REVERSE)
+    const auto mark_traffic_nodes = [&](auto const &traffic_control_node) {
+        if (traffic_control_node.second == TrafficFlowControlNodeDirection::FORWARD ||
+            traffic_control_node.second == TrafficFlowControlNodeDirection::REVERSE)
         {
-            unidirectional_signals.insert({traffic_signal.first, traffic_signal.second});
+            unidirectional_traffic_nodes.insert(
+                {traffic_control_node.first, traffic_control_node.second});
         }
         else
         {
-            BOOST_ASSERT(traffic_signal.second == TrafficLightClass::DIRECTION_ALL);
-            bidirectional_signals.insert(traffic_signal.first);
+            BOOST_ASSERT(traffic_control_node.second == TrafficFlowControlNodeDirection::ALL);
+            bidirectional_traffic_nodes.insert(traffic_control_node.first);
         }
     };
-    std::for_each(external_traffic_signals.begin(), external_traffic_signals.end(), mark_signals);
+    std::for_each(external_nodes.begin(), external_nodes.end(), mark_traffic_nodes);
 
-    // Extract all the segments that lead up to unidirectional traffic signals.
+    // Extract all the segments that lead up to unidirectional traffic flow control nodes.
     const auto set_segments = [&](const size_t way_list_idx, auto const & /*unused*/) {
         const auto node_start_offset =
             used_node_id_list.begin() + way_node_id_offsets[way_list_idx];
@@ -1198,24 +1206,24 @@ ExtractionContainers::ReferencedTrafficSignals ExtractionContainers::IdentifyTra
 
         for (auto node_it = node_start_offset; node_it < node_end_offset; node_it++)
         {
-            const auto sig = unidirectional_signals.find(*node_it);
-            if (sig != unidirectional_signals.end())
+            const auto sig = unidirectional_traffic_nodes.find(*node_it);
+            if (sig != unidirectional_traffic_nodes.end())
             {
-                if (sig->second == TrafficLightClass::DIRECTION_FORWARD)
+                if (sig->second == TrafficFlowControlNodeDirection::FORWARD)
                 {
                     if (node_it != node_start_offset)
                     {
                         // Previous node leads to signal
-                        signal_segments.insert({*node_it, *(node_it - 1)});
+                        node_segments.insert({*node_it, *(node_it - 1)});
                     }
                 }
                 else
                 {
-                    BOOST_ASSERT(sig->second == TrafficLightClass::DIRECTION_REVERSE);
+                    BOOST_ASSERT(sig->second == TrafficFlowControlNodeDirection::REVERSE);
                     if (node_it + 1 != node_end_offset)
                     {
                         // Next node leads to signal
-                        signal_segments.insert({*node_it, *(node_it + 1)});
+                        node_segments.insert({*node_it, *(node_it + 1)});
                     }
                 }
             }
@@ -1224,7 +1232,7 @@ ExtractionContainers::ReferencedTrafficSignals ExtractionContainers::IdentifyTra
     util::for_each_indexed(ways_list.cbegin(), ways_list.cend(), set_segments);
 
     util::for_each_pair(
-        signal_segments, [](const auto pair_a, const auto pair_b) {
+        node_segments, [](const auto pair_a, const auto pair_b) {
             if (pair_a.first == pair_b.first)
             {
                 // If a node is appearing multiple times in this map, then it's ambiguous.
@@ -1240,10 +1248,10 @@ ExtractionContainers::ReferencedTrafficSignals ExtractionContainers::IdentifyTra
             }
         });
 
-    TIMER_STOP(identify_traffic_signals);
-    log << "ok, after " << TIMER_SEC(identify_traffic_signals) << "s";
+    TIMER_STOP(identify_traffic_flow_control_nodes);
+    log << "ok, after " << TIMER_SEC(identify_traffic_flow_control_nodes) << "s";
 
-    return {std::move(bidirectional_signals), std::move(signal_segments)};
+    return {std::move(bidirectional_traffic_nodes), std::move(node_segments)};
 }
 
 void ExtractionContainers::PrepareRestrictions(const ReferencedWays &restriction_ways)